% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_law.R
\docType{data}
\name{law}
\alias{law}
\title{Law School Data}
\format{
A data frame with 82 cases and 2 variables:
\describe{
\item{LSAT}{
Law School Admission Test.
}
\item{GPA}{
Grade Point Average.
}
}
}
\source{
Table 3.2 in Efron and Tibshirani (1993).
}
\usage{
law
}
\description{
The population of measurements (\code{LSAT}, \code{GPA}),
for the universe of 82 American law schools
participating in a large study of admission practices.
Two measurements were made on the entering classes of each school in 1973: \code{LSAT},
the average score for the class on a national law test,
and \code{GPA},
the average undergraduate grade-point average for the class.
}
\examples{
data(law, package = "jeksterslabRdatarepo")
str(law)
head(law)
}
\references{
Efron, B., & Tibshirani, R. J. (1993).
\emph{An introduction to the bootstrap}.
New York, N.Y: Chapman & Hall.
ISBN:9780412042317.
Google:\href{https://books.google.com/books/about/An_Introduction_to_the_Bootstrap.html?id=gLlpIUxRntoC}{gLlpIUxRntoC}.
}
\keyword{datasets}
