% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_lsat.R
\docType{data}
\name{lsat}
\alias{lsat}
\title{Law School Data}
\format{
A matrix with 82 cases and 2 variables:
\describe{
\item{LSAT}{Law School Admission Test.}
\item{GPA}{Grade Point Average.}
}
}
\source{
Efron, B., & Tibshirani, R. J. (1993).
\emph{An introduction to the bootstrap}.
New York, N.Y: Chapman & Hall.
Table 3.2
}
\usage{
lsat
}
\description{
Law School Data
}
\keyword{datasets}
