% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_salary.R
\docType{data}
\name{salary}
\alias{salary}
\title{Salaries and Related Data from 62 Academics}
\format{
A data frame with 62 cases and 5 variables:
\describe{
\item{phd}{
Time since Ph.D.
}
\item{pubs}{
Number of publications.
}
\item{sex}{
Sex. 1 female. 0 male.
}
\item{citations}{
Number of citations.
}
\item{salary}{
Salary in US dollars.
}
}
}
\source{
From Chapter 3 in Cohen et al. (2013).
}
\usage{
salary
}
\description{
Salaries and Related Data from 62 Academics
}
\examples{
data(salary, package = "jeksterslabRdatarepo")
str(salary)
head(salary)
}
\references{
Cohen, J., Cohen, P., West, S. G., & Aiken, L. S. (2013). \emph{Applied Multiple Regression/Correlation Analysis for the Behavioral Sciences}.
Hoboken: Taylor and Francis.
ISBN:9780805822236
Google:\href{https://books.google.com/books/about/Applied_Multiple_Regression_correlation.html?id=v2pGI8RDAPMC}{v2pGI8RDAPMC}
}
\keyword{datasets}
