% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_svd.linreg.R
\docType{data}
\name{svd.linreg}
\alias{svd.linreg}
\title{Singular Value Decomposition in Regression Analysis}
\format{
A list with four elements \code{ya}, \code{Xa}, \code{yb}, and \code{Xb}.
\verb{y*} is a vector of length 8 and \verb{X*} is an 8 by 3 matrix.
\code{ya} and \code{Xa} correspond to Table 1 in Mandel (1982).
\code{yb} and \code{Xb} correspond to Table 4 in Mandel (1982).
}
\source{
Tables 1 and 4 in Mandel (1982)
}
\usage{
svd.linreg
}
\description{
This data set can be used to demostrate the use of
singular value decomposition in regression analysis.
}
\examples{
data(svd.linreg, package = "jeksterslabRdatarepo")
str(svd.linreg)
head(svd.linreg)
}
\references{
Mandel, J. (1982). Use of the Singular Value Decomposition in Regression Analysis.
\emph{The American Statistician}, \emph{36}(1), 15-24.
doi:\href{https://doi.org/10.1080/00031305.1982.10482771}{10.1080/00031305.1982.10482771}.
}
\keyword{datasets}
