% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_svd.R
\docType{data}
\name{svd}
\alias{svd}
\title{Singular Value Decomposition in Regression Analysis}
\format{
A list with two elements \code{y} and \code{X}.
\code{y} is a vector of length 8 and \code{X} is an 8 by 3 matrix.
}
\source{
Table 1 in Mandel (1982)
}
\usage{
svd
}
\description{
This data set can be used to demostrate the use of
singular value decomposition in regression analysis.
}
\examples{
data(svd, package = "jeksterslabRdatarepo")
str(svd)
head(svd)
}
\references{
Mandel, J. (1982). Use of the Singular Value Decomposition in Regression Analysis.
\emph{The American Statistician}, \emph{36}(1), 15-24.
doi:\href{https://doi.org/10.1080/00031305.1982.10482771}{10.1080/00031305.1982.10482771}.
}
\keyword{datasets}
