% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_thirst.R
\docType{data}
\name{thirst}
\alias{thirst}
\title{Temperature on Water Consumption Via Thirst}
\format{
A data frame with 50 cases and 3 variables:
\describe{
\item{temp}{
Room temperature in degrees Fahrenheit.
}
\item{thirst}{
Self-reported thirst at the end of a 2-hour period.
}
\item{water}{
Water consumed during the last 2 hours in deciliters.
}
}
}
\source{
Table 3.1 in MacKinnon (2008)
}
\usage{
thirst
}
\description{
A hypothetical study
of the effects of temperature
on water consumption.
}
\examples{
data(thirst, package = "jeksterslabRdatarepo")
str(thirst)
head(thirst)
}
\references{
MacKinnon, D. P. (2008).
\emph{Introduction to statistical mediation analysis}.
New York: Lawrence Erlbaum Associates.
ISBN:9780805839746.
Google:\href{https://books.google.com/books/about/Introduction_to_Statistical_Mediation_An.html?id=vHUVYKfrdtYC}{vHUVYKfrdtYC}.
}
\keyword{datasets}
