% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wages.matrix.R
\docType{data}
\name{wages.matrix}
\alias{wages.matrix}
\title{Wages and Related Data for 1289 Workers (Matrix Format)}
\format{
A list with two elements:
\describe{
\item{y}{wages}
\item{X}{gender, race, union, education, experiemce, and age}
}
\code{X} has the following elements:
\describe{
\item{gender}{
Gender, 1 for female, 0 for male.
}
\item{race}{
Race, 1 for nonwhite workers and 0 for white workers.
}
\item{union}{
Union status, 1 if in union, 0 otherwise.
}
\item{education}{
Education in years.
}
\item{experience}{
Work experience in years.
}
\item{age}{
Age in years.
}
}
Both \code{y} and \code{X} have 1289 rows.
}
\source{
From Chapter 4 in Gujarati (2019).
}
\usage{
wages.matrix
}
\description{
A random sample of 1,280 workers
from the Cirrent Population Survey (CPS)
for March 1995.
}
\examples{
data(wages.matrix, package = "jeksterslabRdatarepo")
str(wages.matrix)
head(wages.matrix[["y"]])
head(wages.matrix[["X"]])
}
\references{
Gujarati, D. N. (2019). \emph{Linear regression: A mathematical introduction}.
Thousand Oaks: SAGE Publications.
ISBN:9781544336558.
Google:\href{https://books.google.com/books/about/Linear_Regression.html?id=pqxyDwAAQBAJ}{pqxyDwAAQBAJ}
}
\keyword{datasets}
