% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wages.R
\docType{data}
\name{wages}
\alias{wages}
\title{Wages and Related Data for 1289 Workers}
\format{
A data frame with 1289 cases and 7 variables:
\describe{
\item{wages}{
Hourly wages in US dollars.
}
\item{gender}{
Gender, 1 for female, 0 for male.
}
\item{race}{
Race, 1 for nonwhite workers and 0 for white workers.
}
\item{union}{
Union status, 1 if in union, 0 otherwise.
}
\item{education}{
Education in years.
}
\item{experience}{
Work experience in years.
}
\item{age}{
Age in years.
}
}
}
\source{
From Chapter 4 in Gujarati(2019).
}
\usage{
wages
}
\description{
A random sample of 1,280 workers
from the Cirrent Population Survey (CPS)
for March 1995.
}
\examples{
data(wages, package = "jeksterslabRdatarepo")
str(wages)
head(wages)
}
\references{
Gujarati, D. N. (2019). \emph{Linear regression: A mathematical introduction}.
Thousand Oaks: SAGE Publications.
ISBN:9781544336558.
Google:\href{https://books.google.com/books/about/Linear_Regression.html?id=pqxyDwAAQBAJ}{pqxyDwAAQBAJ}
}
\keyword{datasets}
