% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_thirst.R
\docType{data}
\name{thirst}
\alias{thirst}
\title{Temperature on Water Consumption Via Thirst}
\format{
A matrix with 50 cases and 3 variables:
\describe{
\item{temp}{Room temperature in degrees Fahrenheit.}
\item{thirst}{Self-reported thirst at the end of a 2-hour period.}
\item{water}{Water consumed during the last 2 hours in deciliters.}
}
}
\source{
MacKinnon, D. P. (2008).
\emph{Introduction to statistical mediation analysis}.
New York: Lawrence Erlbaum Associates.
Table 3.1
}
\usage{
thirst
}
\description{
Temperature on Water Consumption Via Thirst
}
\keyword{datasets}
